CardPlannerModel/src/uk/co/itstherules/cardplanner/DateCalendarFactory.java
package uk.co.itstherules.cardplanner;

import java.util.Calendar;
import java.util.Date;
import java.util.GregorianCalendar;

public final class DateCalendarFactory {

    private DateCalendarFactory(){}
	
	public static Calendar instance(Date date) {
		Calendar calendar = GregorianCalendar.getInstance();
		calendar.setTimeInMillis(date.getTime());
		calendar.set(Calendar.HOUR_OF_DAY, 0);
		calendar.set(Calendar.MINUTE, 0);
		calendar.set(Calendar.SECOND, 0);
		calendar.set(Calendar.MILLISECOND, 0);
		return calendar;
	}
	
	public static Calendar instance(int date, int month, int year){
		Calendar calendar = GregorianCalendar.getInstance();
		calendar.set(Calendar.DAY_OF_MONTH, date);
		calendar.set(Calendar.MONTH, month);
		calendar.set(Calendar.YEAR, year);
		calendar.set(Calendar.HOUR_OF_DAY, 0);
		calendar.set(Calendar.MINUTE, 0);
		calendar.set(Calendar.SECOND, 0);
		calendar.set(Calendar.MILLISECOND, 0);
		return calendar;
	}
}
